Move test macros to a header
